added JSdocs to all js files
