+++ src/js/api/auth/edit.mjs
@@ -4,6 +4,16 @@ import { API_URL } from "../constants.mjs";
 const action = "/profiles/";
 const method = "PUT";
 
+/**
+ * @async
+ * @module api/auth/edit
+ * @description This function sends a PUT request to the API to edit a profile. It uses the authFetch function to send the request with the access token.
+ * @see module:authFetch
+ * @param {object} profileData
+ * @param {string} name
+ * @returns profileObject from API
+ */
+
 export async function edit(profileData, name) {
   const url = `${API_URL}${action}${name}/media`;
   const body = JSON.stringify(profileData);
