+++ src/js/api/auth/index.mjs
@@ -1,3 +1,11 @@
 export * from "./edit.mjs"
 export * from "./login.mjs"
 export * from "./register.mjs"
+
+/** 
+ * @module auth
+ * @description This module contains all the functions related to authentication.
+ * @see module:auth/edit
+ * @see module:auth/login
+ * @see module:auth/register
+ */
