+++ src/js/api/auth/login.mjs
@@ -5,6 +5,21 @@ import * as profile from "../profile/index.mjs";
 const action = "/auth/login";
 const method = "POST";
 
+/**
+ * @async
+ * @module api/auth/login
+ * @description This function sends a POST request to the API to log in a user.
+ * @param {object} profileData - Login data from the login form.
+ * @returns {void}
+ * @throws {error} An error is thrown if the API call fails.
+ * @example
+ * const profileData = {
+ *  name: "username",
+ * password: "password",
+ * };
+ * login(profileData);
+ */
+
 export async function login(profileData) {
   const url = `${API_URL}${action}`;
   const body = JSON.stringify(profileData);
