+++ src/js/api/posts/index.mjs
@@ -1,4 +1,13 @@
-export * from './create.mjs';
-export * from './get.mjs';
-export * from './delete.mjs';
-export * from './update.mjs';
+export * from "./create.mjs";
+export * from "./get.mjs";
+export * from "./delete.mjs";
+export * from "./update.mjs";
+
+/**
+ * @module posts
+ * @description This module contains all the functions related to posts.
+ * @see module:posts/create
+ * @see module:posts/get
+ * @see module:posts/delete
+ * @see module:posts/update
+ */
