+++ src/js/api/profile/index.mjs
@@ -1,2 +1,9 @@
 export * from "./get.mjs";
 export * from "./follow.mjs";
+
+/**
+ * @module profile
+ * @description This module contains all the functions related to profiles.
+ * @see module:profile/get
+ * @see module:profile/follow
+ */
