+++ src/js/display/index.mjs
@@ -3,3 +3,13 @@ export * from "./profile.mjs";
 export * from "./posts/post.mjs";
 export * from "./posts/posts.mjs";
 export * from "./posts/profilePosts.mjs";
+
+/**
+ * @module display
+ * @description This module contains all the functions related to displaying content.
+ * @see module:display/navBar
+ * @see module:display/profile
+ * @see module:display/posts/post
+ * @see module:display/posts/posts
+ * @see module:display/posts/profilePosts
+ */
