+++ src/js/handlers/navigation/index.mjs
@@ -1,3 +1,11 @@
 export * from "./loggedInChecker.mjs";
 export * from "./loggedInNavBar.mjs";
 export * from "./navDropDown.mjs";
+
+/**
+ * @module navigation
+ * @description This module contains all the functions related to navigation.
+ * @see module:navigation/loggedInChecker
+ * @see module:navigation/loggedInNavBar
+ * @see module:navigation/navDropDown
+ */
