+++ src/js/handlers/navigation/loggedInNavBar.mjs
@@ -2,6 +2,15 @@ import * as storage from "../../storage/index.mjs";
 import * as display from "../../display/index.mjs";
 import * as URL from "../../url/index.mjs";
 
+/**
+ * @module handlers/navigation/loggedInNavBar
+ * @description This module displays the navigation bar for logged in users. It uses the storage module to get the profile information.
+ * @see module:storage
+ * @see module:display
+ * @see module:url
+ * @returns {void}
+ */
+
 export function setLoggedInNavBarListener() {
   const path = URL.getPath();
   const navList = document.querySelector("#navList");
