+++ src/js/handlers/posts/index.mjs
@@ -1,3 +1,11 @@
 export * from "./create.mjs";
-export * from "./delete.mjs";
+export * from "./searchAndFilter.mjs";
 export * from "./update.mjs";
+
+/**
+ * @module handlers/posts
+ * @description This module contains all the functions related to posts.
+ * @see module:handlers/posts/create
+ * @see module:handlers/posts/searchAndFilter
+ * @see module:handlers/posts/update
+ */
