+++ src/js/handlers/profile/index.mjs
@@ -3,3 +3,13 @@ export * from "./follow.mjs";
 export * from "./login.mjs";
 export * from "./logout.mjs";
 export * from "./register.mjs";
+
+/**
+ * @module handlers/profile
+ * @description This module contains all the handlers functions related to the profile.
+ * @see module:handlers/profile/edit
+ * @see module:handlers/profile/follow
+ * @see module:handlers/profile/login
+ * @see module:handlers/profile/logout
+ * @see module:handlers/profile/register
+ */
